osd/PrimaryLogPG.cc: fix set-alloc-hint and migration conflict
client(rbd client for exmaple) sometimes send set-alloc-hint + write
ops in one message. As client don't know object's fast tag, the
set-alloc-hint flag is 0. But in migration logic, this is a demote
hint. When a set-alloc-hint op received on osd side, it will be
processed like the following step:

1. object tag is slow
2. after checked the hit set, this object should be promoted,
   and object's tag is set to fast
3. process set-alloc-hint op:
   op.flags == 0 and object's flag is fast now, which will
   trigger off a demote op(set object's flag to slow)

The final result is the promotion op is canceled which is not
what we expect.

So we only treat single set-alloc-hint op as migration hint op,
all other set-alloc-hint which mixed with other ops will be
treated as normal hint op.

Signed-off-by: Yang Honggang <yanghonggang@umcloud.com>