osd/PrimaryLogPG.cc: demote before extend a fast object when tier is full
when tier agent is in FULL mode, demote the object to slow device before
extend it(writefull/append/write more). For a pinned object, an auto unpin
will also be taken before the extend ops.

And a osd_tier_inject_cache_mode_full config option is added for test
purpose.

Signed-off-by: Yang Honggang <yanghonggang@umcloud.com>